feat(core): carve @plannotator/core — move pure modules, extract node-bound types, shim shared (Phase 7 step 1)
