+++ packages/core/agent-terminal.test.ts
