+++ packages/core/config-types.ts
@@ -1,0 +1,18 @@
+export type DefaultDiffType = 'since-base' | 'uncommitted' | 'unstaged' | 'staged' | 'merge-base' | 'all';
+export type DiffLineBgIntensity = 'subtle' | 'normal' | 'strong';
+
+export interface DiffOptions {
+  diffStyle?: 'split' | 'unified';
+  overflow?: 'scroll' | 'wrap';
+  diffIndicators?: 'bars' | 'classic' | 'none';
+  lineDiffType?: 'word-alt' | 'word' | 'char' | 'none';
+  showLineNumbers?: boolean;
+  showDiffBackground?: boolean;
+  fontFamily?: string;
+  fontSize?: string;
+  tabSize?: number;
+  hideWhitespace?: boolean;
+  expandUnchanged?: boolean;
+  defaultDiffType?: DefaultDiffType;
+  lineBgIntensity?: DiffLineBgIntensity;
+}
