+++ packages/core/extract-code-paths.test.ts
