+++ packages/core/goal-setup.test.ts
