+++ packages/core/index.ts
@@ -1,0 +1,2 @@
+export * from './ai-context';
+export type { EditorAnnotation } from './types';
