+++ packages/core/storage-types.ts
@@ -1,0 +1,8 @@
+export interface ArchivedPlan {
+  filename: string;
+  title: string;
+  date: string;
+  timestamp: string;  // ISO string from file mtime
+  status: "approved" | "denied" | "unknown";
+  size: number;
+}
